refactor: move `CollectionFiles` to `libAnki` module
Removes a cyclic dependency: libAnki <-> AnkiDroid

Issue 18015
