+++ AnkiDroid/src/androidTest/java/com/ichi2/anki/tests/Shared.kt
@@ -18,10 +18,10 @@ package com.ichi2.anki.tests
 import android.annotation.SuppressLint
 import android.content.Context
 import androidx.test.espresso.matcher.ViewMatchers.assertThat
-import com.ichi2.anki.CollectionFiles
 import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.compat.CompatHelper
 import com.ichi2.libanki.Collection
+import com.ichi2.libanki.CollectionFiles
 import com.ichi2.libanki.Storage
 import org.hamcrest.CoreMatchers.containsString
 import org.hamcrest.CoreMatchers.not
