+++ AnkiDroid/src/main/java/com/ichi2/libanki/CollectionFiles.kt
@@ -1,0 +1,31 @@
+/*
+ *  Copyright (c) 2025 Arthur Milchior <arthur@milchior.fr>
+ *  Copyright (c) 2025 David Allison <davidallisongithub@gmail.com>
+ *
+ *  This program is free software; you can redistribute it and/or modify it under
+ *  the terms of the GNU General Public License as published by the Free Software
+ *  Foundation; either version 3 of the License, or (at your option) any later
+ *  version.
+ *
+ *  This program is distributed in the hope that it will be useful, but WITHOUT ANY
+ *  WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+ *  PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with
+ *  this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.ichi2.libanki
+
+import com.ichi2.libanki.utils.NotInLibAnki
+import java.io.File
+
+@NotInLibAnki
+class CollectionFiles(
+    folderPath: File,
+    val collectionName: String = "collection",
+) {
+    val colDb = File(folderPath, "$collectionName.anki2")
+    val mediaFolder = File(folderPath, "$collectionName.media")
+    val mediaDb = File(folderPath, "$collectionName.media.db")
+}
