+++ AnkiDroid/src/main/java/com/ichi2/libanki/Storage.kt
@@ -15,7 +15,6 @@
  */
 package com.ichi2.libanki
 
-import com.ichi2.anki.CollectionFiles
 import com.ichi2.anki.common.time.Time
 import com.ichi2.anki.common.time.TimeManager.time
 import com.ichi2.anki.common.time.getDayStart
